pyproject.toml
[project]
name = "imagai"
version = "0.1.0"
description = "A CLI tool to generate images using various AI APIs (OpenAI, DALL-E, Stable Diffusion, Gemini, Imagen via OpenAI-compatible interfaces)."
authors = [
    { name = "Chophe", email = "chophe@yahoo.com" }
]
dependencies = [
    "typer[all]>=0.9.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pillow>=10.0.0",
    "openai>=1.0.0",
    "rich>=13.0.0",
]
readme = "README.md"
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.0.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/imagai"]

[project.scripts]
imagai = "imagai.cli:app"
